feat: new schema file organization
