+++ packages/schemas/src/index.ts
@@ -1,1 +1,3 @@
+export * from './lib/ILitNodeClient';
+export * from './lib/models';
 export * from './lib/schemas';
